% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.means.R
\name{comp.means.default}
\alias{comp.means.default}
\title{comp.means.default}
\usage{
\method{comp.means}{default}(
  x,
  ...,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  DEBUG = FALSE,
  paired = FALSE,
  show.descriptives = TRUE,
  show.variance.test = TRUE,
  show.comp.test = TRUE,
  show.df = TRUE,
  show.stat.name = TRUE,
  show.stat.value = TRUE,
  show.stat.ci = TRUE,
  show.p.value.exact = TRUE,
  show.p.value = TRUE,
  show.p.symbols = TRUE,
  show.estimate = TRUE,
  p.sig = 0.05,
  p.sig.small = 0.01,
  p.sig.very.small = 0.001,
  ci = 0.95,
  decimals = 2
)
}
\description{
comp.means.default
}
