% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{comp.media}
\alias{comp.media}
\title{comp.media}
\usage{
comp.media(
  ...,
  by = NULL,
  decimals = 2,
  DEBUG = FALSE,
  DEBUG.FORMA = FALSE,
  DEBUG.CALL = FALSE,
  show.vars = TRUE,
  show.by = TRUE,
  show.groups = TRUE,
  show.n.valid = TRUE,
  show.n.missing = TRUE,
  show.min = TRUE,
  show.max = TRUE,
  show.mean = TRUE,
  show.sd = TRUE,
  show.median = TRUE,
  show.IRQ = TRUE,
  show.se = TRUE,
  show.ci.upper = TRUE,
  show.ci.lower = TRUE,
  show.p.norm = TRUE,
  show.p.norm.exact = FALSE,
  show.nor.test = TRUE,
  show.is.normal = TRUE,
  show.interpretation = FALSE,
  lang = "es",
  show.global = TRUE,
  p.sig = 0.05,
  p.sig.small = 0.01,
  p.sig.very.small = 0.001,
  ci = 0.95,
  show.post.hoc = TRUE,
  show.desc = TRUE,
  paired = T
)
}
\arguments{
\item{by}{default value: NULL}

\item{decimals}{default value: 2}

\item{DEBUG}{default value: FALSE}

\item{DEBUG.FORMA}{default value: FALSE}

\item{DEBUG.CALL}{default value: FALSE}

\item{show.vars}{default value: TRUE}

\item{show.by}{default value: TRUE}

\item{show.groups}{default value: TRUE}

\item{show.n.valid}{default value: TRUE}

\item{show.n.missing}{default value: TRUE}

\item{show.min}{default value: TRUE}

\item{show.max}{default value: TRUE}

\item{show.mean}{default value: TRUE}

\item{show.sd}{default value: TRUE}

\item{show.median}{default value: TRUE}

\item{show.IRQ}{default value: TRUE}

\item{show.se}{default value: TRUE}

\item{show.ci.upper}{default value: TRUE}

\item{show.ci.lower}{default value: TRUE}

\item{show.p.norm}{default value: TRUE}

\item{show.p.norm.exact}{default value: FALSE}

\item{show.nor.test}{default value: TRUE}

\item{show.is.normal}{default value: TRUE}

\item{show.interpretation}{default value: FALSE}

\item{lang}{default value: "es"}

\item{show.global}{default value: TRUE}

\item{p.sig}{default value: 0.05}

\item{p.sig.small}{default value: 0.01}

\item{p.sig.very.small}{default value: 0.001}

\item{ci}{default value: 0.95}

\item{show.post.hoc}{default value: TRUE}

\item{show.desc}{default value: TRUE}

\item{paired}{default value: TRU}
}
\value{

}
\description{
comp.media
}
