% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_means.R
\name{comp_means.data.frame}
\alias{comp_means.data.frame}
\title{comp_means.data.frame}
\usage{
\method{comp_means}{data.frame}(
  x,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  ci = 0.95,
  alternative = "two.sided",
  lang = "es",
  decimals = 2,
  method = "auto",
  paired = FALSE,
  p.sig = 0.05,
  p.sig.small = 0.01,
  p.sig.very.small = 0.001,
  show.descriptives = TRUE,
  show.variance = TRUE,
  DEBUG = FALSE,
  stop.on.error = TRUE,
  show.error = TRUE,
  ...
)
}
\description{
comp_means.data.frame
}
