% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_means.R
\name{comp_means}
\alias{comp_means}
\title{comp_means}
\usage{
comp_means(
  x,
  xname = NULL,
  by = NULL,
  byname = NULL,
  ci = 0.95,
  alternative = "two.sided",
  decimals = 2,
  method = "auto",
  paired = FALSE,
  p.sig = 0.05,
  p.sig.small = 0.01,
  p.sig.very.small = 0.001,
  show.descriptives = TRUE,
  show.variance = TRUE,
  stop.on.error = TRUE,
  show.error = TRUE,
  lang = "es",
  DEBUG = FALSE,
  ...
)
}
\description{
comp_means
}
