% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.means.R
\name{.comp.means.2.groups}
\alias{.comp.means.2.groups}
\title{comp.means.numeric}
\usage{
.comp.means.2.groups(
  x,
  y = NULL,
  ...,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  DEBUG = FALSE,
  paired = FALSE,
  p.sig = 0.05,
  p.sig.small = 0.01,
  p.sig.very.small = 0.001,
  ci = 0.95
)
}
\description{
comp.means.numeric
}
