% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.means.R
\name{.comp.means.form.result}
\alias{.comp.means.form.result}
\title{.comp.means.form.result}
\usage{
.comp.means.form.result(
  result,
  res = NULL,
  DEBUG = FALSE,
  show.descriptives = TRUE,
  show.variance.test = TRUE,
  show.comp.test = TRUE,
  show.df = TRUE,
  show.stat.name = TRUE,
  show.stat.value = TRUE,
  show.stat.ci = TRUE,
  show.p.value.exact = TRUE,
  show.p.value = TRUE,
  show.p.symbols = TRUE,
  show.estimate = TRUE
)
}
\description{
Merges results and creates the proper attributes
}
\details{
If the `res` object is present it merges it with `result`
If not pressent it just applies all the attributes to `result`
}
