% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.msgs.R, R/utils.R
\name{.error.msg}
\alias{.error.msg}
\title{means.default}
\usage{
.error.msg(er = "DEFAULT", lang = "es", as.text = TRUE)

.error.msg(er = "DEFAULT", lang = "es", as.text = TRUE)
}
\description{
means.default

means.default
}
