% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.t.tests.R
\name{.t.test.results}
\alias{.t.test.results}
\title{.t.test.results}
\usage{
.t.test.results(
  test,
  x.name = NULL,
  y.name = NULL,
  y.levels = NULL,
  test.name = NULL,
  alternative = NULL,
  decimals = 4
)
}
\description{
.t.test.results
}
