% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.wilcoxon.R
\name{.wilcoxon.results}
\alias{.wilcoxon.results}
\title{.t.test.results}
\usage{
.wilcoxon.results(
  test,
  x.name = NULL,
  y.name = NULL,
  y.levels = NULL,
  testname = NULL,
  alternative = NULL,
  decimals = 2
)
}
\description{
.t.test.results
}
