% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{ECM}
\alias{ECM}
\title{An Error Correction Model Function}
\usage{
ECM(formula, serie)
}
\arguments{
\item{formula}{A regression formula.}

\item{serie}{A dataframe or ts object with the series used by the formula provided.}
}
\description{
This function transforms your OLS coefficients into error correction coefficients. It also returns a list with the p-values for stargazer/texreg packages.
}
\examples{
ECM(y ~ x + z, df)
}
\author{
Fernando Teixeira
}
\keyword{Correction}
\keyword{ECM}
\keyword{Error}
\keyword{Model,}
