% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo.R
\name{euler}
\alias{euler}
\title{A Cat Function}
\usage{
euler(X0, mu, sigma, Dt, t, T, N, plt = F)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

