% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_call.R
\name{euler_call}
\alias{euler_call}
\title{A Cat Function}
\usage{
euler_call(X0, mu, sigma, Dt = 0.1, t, T, N = 100, K = 0, plt = FALSE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

