% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_adf.R
\name{func_adf}
\alias{func_adf}
\title{A Cat Function}
\usage{
func_adf(data, col.init, col.fin, lags = 24, select.lags = "AIC",
  mymfrow = c(5, 3), mymar = c(3, 2, 1.5, 1), myoma = c(0, 0, 2, 0),
  dif = "y", res = "y", precision = 2, plt = F)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

