% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.normality.R
\name{is.normal}
\alias{is.normal}
\title{is.normal}
\usage{
is.normal(x, p.sig = 0.05)
}
\arguments{
\item{x}{numeric vector}

\item{p.sig}{(default=0.05) value under wich p has to be to reject normality}
}
\value{
(bool) TRUE if the vector is normal, FALSE in any other case
}
\description{
Evaluates if a vector 'x' is follows a normal distribution or not
}
\examples{

is.norma(mtcars$mpg)

}
