% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{means.data.frame}
\alias{means.data.frame}
\title{means.data.frame}
\usage{
\method{means}{data.frame}(
  x,
  ...,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  decimals = 2,
  ci = 0.95,
  DEBUG = FALSE,
  stop.on.error = TRUE,
  show.errors = TRUE,
  lang = "es"
)
}
\description{
means.data.frame
}
