% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{means.numeric}
\alias{means.numeric}
\title{means.numeric}
\usage{
\method{means}{numeric}(
  x,
  ...,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  decimals = 2,
  DEBUG = FALSE,
  show.vars = TRUE,
  show.by = TRUE,
  show.groups = TRUE,
  show.n.valid = TRUE,
  show.n.missing = TRUE,
  show.min = TRUE,
  show.max = TRUE,
  show.mean = TRUE,
  show.sd = TRUE,
  show.median = TRUE,
  show.IRQ = TRUE,
  show.se = FALSE,
  show.ci.upper = FALSE,
  show.ci.lower = FALSE,
  show.p.norm = TRUE,
  show.p.norm.exact = FALSE,
  show.nor.test = TRUE,
  show.is.normal = TRUE,
  show.interpretation = FALSE,
  lang = "es",
  show.global = TRUE,
  p.sig = 0.05,
  p.sig.small = 0.01,
  p.sig.very.small = 0.001,
  ci = 0.95,
  comp = FALSE,
  show.post.hoc = TRUE,
  show.desc = TRUE,
  paired = T,
  errors.as.text = FALSE
)
}
\description{
means.numeric
}
