% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{means}
\alias{means}
\title{#' @rdname media
#' @export
m <- media}
\usage{
means(
  x,
  ...,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  decimals = 2,
  ci = 0.95,
  DEBUG = FALSE,
  stop.on.error = TRUE,
  show.errors = TRUE,
  lang = "es"
)
}
\description{
means
}
