% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{media}
\alias{media}
\alias{m}
\title{media}
\usage{
media(
  x,
  ...,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  decimals = 2,
  show.var.name = TRUE,
  show.group.var = TRUE,
  show.group = TRUE,
  show.n.valid = TRUE,
  show.n.missing = TRUE,
  show.min = TRUE,
  show.max = TRUE,
  show.mean = TRUE,
  show.sd = TRUE,
  show.median = TRUE,
  show.IQR = TRUE,
  show.se = FALSE,
  show.ci.upper = FALSE,
  show.ci.lower = FALSE,
  show.p.norm = TRUE,
  show.p.norm.exact = FALSE,
  show.nor.test = TRUE,
  show.is.normal = TRUE,
  show.interpretation = FALSE,
  lang = "es",
  show.global = TRUE,
  show.help = FALSE,
  p.sig = 0.05,
  p.sig.small = 0.01,
  p.sig.very.small = 0.001,
  ci = 0.95,
  comp = FALSE,
  show.post.hoc = TRUE,
  show.desc = TRUE,
  paired = T,
  stop.on.error = FALSE,
  show.error = TRUE,
  DEBUG = FALSE
)

m(
  x,
  ...,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  decimals = 2,
  show.var.name = TRUE,
  show.group.var = TRUE,
  show.group = TRUE,
  show.n.valid = TRUE,
  show.n.missing = TRUE,
  show.min = TRUE,
  show.max = TRUE,
  show.mean = TRUE,
  show.sd = TRUE,
  show.median = TRUE,
  show.IQR = TRUE,
  show.se = FALSE,
  show.ci.upper = FALSE,
  show.ci.lower = FALSE,
  show.p.norm = TRUE,
  show.p.norm.exact = FALSE,
  show.nor.test = TRUE,
  show.is.normal = TRUE,
  show.interpretation = FALSE,
  lang = "es",
  show.global = TRUE,
  show.help = FALSE,
  p.sig = 0.05,
  p.sig.small = 0.01,
  p.sig.very.small = 0.001,
  ci = 0.95,
  comp = FALSE,
  show.post.hoc = TRUE,
  show.desc = TRUE,
  paired = T,
  stop.on.error = FALSE,
  show.error = TRUE,
  DEBUG = FALSE
)
}
\arguments{
\item{x}{data.frame or vector with data}

\item{...}{if data.frame was specified variables can be used here to select columns}

\item{xname}{optional. Name for 'x' variable}

\item{by}{optional. Variable to group 'x' by and can be:

                              - a vector with the same length as 'x'
                              - a data.frame
                              - if 'x' is a data.frame it can be a string (or vector of strings) with column, names present in 'x'}

\item{byname}{optional. Name for 'by' variable}

\item{decimals}{optional. Default = 2. Number of decimals to show.}

\item{show.var.name}{optional. Default = TRUE. Show var.name column on print}

\item{show.group.var}{optional. Default = TRUE. Show group.var column on print}

\item{show.group}{optional. Default = TRUE. Show group column on print}

\item{show.n.valid}{optional. Default = TRUE. Show n.valid column on print}

\item{show.n.missing}{optional. Default = TRUE. Show n.missing column on print}

\item{show.min}{optional. Default = TRUE. Show min column on print}

\item{show.max}{optional. Default = TRUE. Show max column on print}

\item{show.mean}{optional. Default = TRUE. Show mean column on print}

\item{show.sd}{optional. Default = TRUE. Show sd column on print}

\item{show.median}{optional. Default = TRUE. Show median column on print}

\item{show.IQR}{optional. Default = TRUE. Show IQR column on print}

\item{show.se}{optional. Default = TRUE. Show se column on print}

\item{show.ci.upper}{optional. Default = TRUE. Show ci.upper column on print}

\item{show.ci.lower}{optional. Default = TRUE. Show ci.lower column on print}

\item{show.p.norm}{optional. Default = TRUE. Show p.norm column on print}

\item{show.p.norm.exact}{optional. Default = TRUE. Show p.norm.exact column on print}

\item{show.nor.test}{optional. Default = TRUE. Show nor.test column on print}

\item{show.is.normal}{optional. Default = TRUE. Show is.normal column on print}

\item{show.interpretation}{optional. Default = FALSE Show a text with an EXAMPLE
to show how the data can be interpreted. This is by no
means a thorough or complete interpretation. Just a guide.}

\item{lang}{}

\item{show.global}{optional. Default = TRUE. If there is a 'by' variable defined this will
prompt the printing of the variable means WITHOUT agrupation too}

\item{p.sig}{optional. Default = 0.05. Value below wich a p.value is considered 'statistically significant'}

\item{p.sig.small}{optional. Default = 0.01. Value below wich a p.value is considered 'statistically significant' and 'small'}

\item{p.sig.very.small}{optional. Value below wich a p.value is considered 'statistically significant' and 'very small'}

\item{ci}{optional. Default = 0.95. Value for the Confidence Interval}

\item{DEBUG}{}
}
\description{
This function will give a data.frame with the necessary statistics to describe
a numeric variable, including selecting the appropriate test to check if the
variable follows a normal distribution and offering the possibilitiy of a brief
interpretation paragraph to show how to describe the variable and a detailed
description of each column in the resulting data.base
}
