% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milstein.R
\name{milstein}
\alias{milstein}
\title{Milstein Function}
\usage{
milstein(X0, mu, sigma, Dt, t, T, N, plt = F)
}
\arguments{
\item{X0}{Initial value.}

\item{mu}{Drift coeficient.}

\item{sigma}{Diffusion coefficient.}

\item{Dt}{Step size.}

\item{t}{Initial time period.}

\item{T}{Final time period.}

\item{N}{Number of simulations.}

\item{plt}{Plot? Defaults to FALSE.}
}
\description{
This function allows you to simulate Milstein method.
}
\examples{
milstein(307.65, 0.75, 0.3, 0.001, 0, 1, 10000)


}
\author{
Fernando Teixeira
}
\keyword{method}
\keyword{milstein}

