% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.means.R
\name{print.feR.comp.mean.total}
\alias{print.feR.comp.mean.total}
\title{print.feR.comp.mean.total}
\usage{
\method{print}{feR.comp.mean.total}(x)
}
\description{
Esta función deberá controlar si se imprimen los descriptivos y formato y esas cosas
}
