% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.feR.means.R
\name{print.feR.means}
\alias{print.feR.means}
\title{print.feR.means}
\usage{
\method{print}{feR.means}(x)
}
\arguments{
\item{x}{feR.means object}
}
\description{
Prints a feR.means objecto with the desired output as indicated in the object
}
\details{
Some of the attributes that can be tweaked are:
- DECIMALS               number of decimals to show
- LANG                   languaje of the text
- P.SIG                  level for witch p values will be found to be 'significant'
- P.SIG.SMALL            level for witch p values will be found to be 'significant' and 'small'
- P.SIG.VERY.SMALL       level for witch p values will be found to be 'significant' and 'very small'
- CI                     confidence interval
- SHOW.VAR.NAME
- SHOW.GROUP.VAR
- SHOW.GROUP
- SHOW.N.VALID
- SHOW.N.MISSING
- SHOW.MIN
- SHOW.MAX
- SHOW.MEAN
- SHOW.SD
- SHOW.MEDIAN
- SHOW.IQR
- SHOW.SE
- SHOW.CI.UPPER
- SHOW.CI.LOWER
- SHOW.P.NORM
- SHOW.P.NORM.EXACT
- SHOW.NOR.TEST
- SHOW.IS.NORMAL
- SHOW.GLOBAL
}
