% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.t.tests.R
\name{t_student}
\alias{t_student}
\title{t_student}
\usage{
t_student(
  x,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  ci = 0.95,
  alternative = "two.sided",
  stop.on.error = FALSE,
  lang = "es",
  decimals = 2
)
}
\description{
t_student
}
