% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.t.tests.R
\name{t_test}
\alias{t_test}
\title{t_test}
\usage{
t_test(
  x,
  x.name = feR:::.var.name(deparse(substitute(x))),
  y = NULL,
  y.name = feR:::.var.name(deparse(substitute(y))),
  ci = 0.95,
  alternative = "two.sided",
  stop.on.error = FALSE,
  lang = "es",
  decimals = 4
)
}
\description{
t_test
}
