% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.tests.R
\name{t_test.student}
\alias{t_test.student}
\title{t_test.student}
\usage{
t_test.student(
  x,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  ci = 0.95,
  alternative = "two.sided",
  stop.on.error = FALSE,
  lang = "es"
)
}
\description{
t_test.student
}
