% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.tests.R
\name{t_test.welch}
\alias{t_test.welch}
\title{=============================================
ORDENANDO COSAS
=============================================
- Estas funciones deberían controlar sus errores básicos
- Devolver un data.frame con la información
- Tener un print que saque la información adecuadamente
- Tener su set de pruebas
- USAR FORMULAS}
\usage{
t_test.welch(
  x,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  ci = 0.95,
  decimals = 2,
  stop.on.error = FALSE
)
}
\description{
Una vez que esté esto resuelto usaré este sistema para ampliarlo a todo
- Primero archivo con las pruebas matemáticas y sus soluciones
- Despues muestra de resultados y ampliación de información (otro archivo)
- Por último un wrapper que permita discriminar qué función usar y agregue todo
    - Este wrapper será el que determine:
          - qué prueba usar
          - si se muestran todas las columnas o no
          - si se muestra interpretación o no
          - gestión más sofisticada de errores
}
\details{
PERO LO PRIMERO es hacer que todo esto funcione suave y pueda tener toda la estadística en orden
t_test.welch
}
