% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste_adf2.R
\name{teste_adf2}
\alias{teste_adf2}
\title{A Cat Function}
\usage{
teste_adf2(serie, lag_max)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
adf2 function
This function allows you to express your love of cats.
}

