% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste_dfgls.R
\name{teste_dfgls}
\alias{teste_dfgls}
\title{dfgls function}
\usage{
teste_dfgls(serie, modelo = c("trend", "constant"), lag_max)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}

