% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste_estacionariedade.R
\name{teste_estacionariedade}
\alias{teste_estacionariedade}
\title{Stationarity Test}
\usage{
teste_estacionariedade(sondagens, financeiras, periodo, variacao)
}
\arguments{
\item{sondagens}{matrix with data including a date column.}

\item{financeiras}{matrix with data including a date column.}

\item{periodo}{Barometer period in the format as.character(seq(as.Date("aaaa-mm-dd"), as.Date("aaaa-mm-dd"), by = "month")).}

\item{variacao}{TRUE or FALSE parameter indicating if the series have negative numbers or not.}
}
\description{
This function compiles a number of different stationarity tests
}
\examples{
teste_estacionariedade()
}
\author{
Ingrid Oliveira
}
\keyword{stationarity}
\keyword{test}
