% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste_kpss.R
\name{teste_kpss}
\alias{teste_kpss}
\title{kpss function}
\usage{
teste_kpss(serie)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

