% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste_kpss2.R
\name{teste_kpss2}
\alias{teste_kpss2}
\title{kpss2 function}
\usage{
teste_kpss2(serie, modelo = c("tau", "mu"))
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\keyword{cats}

