% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toda.R
\name{toda}
\alias{toda}
\title{ADF Function}
\usage{
toda(x, m, plt = F, ic = 1, pval = 0.05)
}
\arguments{
\item{x}{Dataframe or tibble.}

\item{m}{Series maximum integration order.}

\item{plt}{Plot? Defaults to FALSE.}

\item{ic}{Information criteria, from 1 to 4 (AIC, HQ, SC, FPE). Defaults to 1 (AIC).}
}
\description{
This function allows you to run an ADF test through several contiguous columns of a dataframe using urca package.
}
\examples{
toda(x, m = 2, plt = F, ic = 1, pval=0.05)
}
\author{
Fernando Teixeira
}
\keyword{Toda-Yamamoto}
