% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.test.R
\name{variance.test.numeric}
\alias{variance.test.numeric}
\title{variance.test.numeric}
\usage{
\method{variance.test}{numeric}(
  x,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  p.sig = 0.05,
  p.sig.small = 0.01,
  p.sig.very.small = 0.001,
  ci = 0.95,
  decimals = 2,
  DEBUG = FALSE,
  lang = "es",
  stop.on.error = TRUE
)
}
\description{
variance.test.numeric
}
