% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.test.R
\name{variance.test}
\alias{variance.test}
\title{TODO:}
\usage{
variance.test(
  x,
  ...,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  p.sig = 0.05,
  p.sig.small = 0.01,
  p.sig.very.small = 0.001,
  ci = 0.95,
  decimals = 2,
  DEBUG = FALSE,
  show.descriptives = TRUE,
  show.variance.test = TRUE,
  show.comp.test = TRUE,
  show.df = TRUE,
  show.stat.name = TRUE,
  show.stat.value = TRUE,
  show.stat.ci = TRUE,
  show.p.value.exact = TRUE,
  show.p.value = TRUE,
  show.p.symbols = TRUE,
  show.estimate = TRUE,
  lang = "es",
  stop.on.error = TRUE,
  show.error = TRUE
)
}
\description{
Implementar el stop.on.error
Implementar lang
}
\details{
Redondeo y eliminación de información inutil en print, pero no en data.frame

variance.test
}
