% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.tests.R
\name{welch_test}
\alias{welch_test}
\title{welch_test}
\usage{
welch_test(
  x,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  ci = 0.95,
  alternative = "two.sided",
  stop.on.error = FALSE,
  lang = "es",
  decimals = 2
)
}
\description{
welch_test
}
