% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.wilcoxon.R
\name{wilcoxon_test}
\alias{wilcoxon_test}
\title{wilcoxon_test}
\usage{
wilcoxon_test(
  x,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  ci = 0.95,
  alternative = "two.sided",
  stop.on.error = TRUE,
  show.error = TRUE,
  lang = "es",
  decimals = 2
)
}
\description{
rstatix::wilcox_test(ToothGrowth ,len ~ supp)
coin::wilcox_test(ToothGrowth$len ~ ToothGrowth$supp, conf.int = T)
wilcox.test(ToothGrowth$len ~ ToothGrowth$supp, conf.int = T, exact = T)
}
