% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.means.R
\name{comp.means}
\alias{comp.means}
\title{comp.means}
\usage{
comp.means(
  x,
  y = NULL,
  ...,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  paired = FALSE,
  decimals = 2
)
}
\description{
comp.means
}
