% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.t.tests.R
\name{.t.test.results}
\alias{.t.test.results}
\title{.t.test.results}
\usage{
.t.test.results(
  test,
  xname = NULL,
  byname = NULL,
  by.levels = NULL,
  testname = NULL,
  alternative = NULL,
  decimals = 2
)
}
\description{
.t.test.results
}
