% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.normality.R
\name{normal.test}
\alias{normal.test}
\title{normal.test}
\usage{
normal.test(
  x,
  n.valid = NULL,
  decimals = 2,
  p.sig = 0.05,
  p.sig.small = 0.01,
  p.sig.very.small = 0.001,
  stop.on.error = TRUE,
  show.error = TRUE
)
}
\description{
normal.test
}
