% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.tests.R
\name{t_test.welch}
\alias{t_test.welch}
\title{t_test.welch}
\usage{
t_test.welch(
  x,
  xname = feR:::.var.name(deparse(substitute(x))),
  by = NULL,
  byname = feR:::.var.name(deparse(substitute(by))),
  ci = 0.95,
  alternative = "two.sided",
  stop.on.error = FALSE,
  lang = "es"
)
}
\description{
t_test.welch
}
