% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste_estacionariedade.R
\name{teste_estacionariedade}
\alias{teste_estacionariedade}
\title{A Cat Function}
\usage{
teste_estacionariedade(sondagens, financeiras, periodo, variacao)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\author{
Ingrid Luquett
}
\keyword{cats}

