% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.t.tests.R
\name{welch_test}
\alias{welch_test}
\title{welch_test}
\usage{
welch_test(
  x,
  x.name = feR:::.var.name(deparse(substitute(x))),
  y = NULL,
  y.name = feR:::.var.name(deparse(substitute(y))),
  p.sig = 0.05,
  alternative = "two.sided",
  stop.on.error = FALSE,
  lang = "es",
  decimals = 4
)
}
\arguments{
\item{x}{DESCRIPTION.}

\item{x.name}{DESCRIPTION.}

\item{y}{DESCRIPTION.}

\item{y.name}{DESCRIPTION.}

\item{alternative}{DESCRIPTION.}

\item{stop.on.error}{DESCRIPTION.}

\item{lang}{DESCRIPTION.}

\item{decimals}{DESCRIPTION.}

\item{ci}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_TITLE
}
\details{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
